requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
pyyaml>=6.0

# Language models and data processing
aiohttp>=3.8.4
numpy>=1.24.3
pandas>=2.0.0
langchain>=0.0.201
jinja2>=3.1.2

# Vector database connections
qdrant-client>=1.2.0
faiss-cpu>=1.7.4

# Infrastructure tools
python-terraform>=0.10.1
ansible-runner>=2.3.2

# Testing and development
pytest>=7.3.1
pytest-asyncio>=0.21.0
black>=23.3.0
flake8>=6.0.0

# Utilities
tqdm>=4.65.0
rich>=13.3.5